apps/pokemonevcalc_app/serializers.py
from rest_framework import serializers
from .models import Move, Pokemon, Items

class MoveSerializer(serializers.ModelSerializer):
    class Meta:
        model = Move
        fields = ['id', 'identifier', 'power', 'damage_class_id', 'type']


class PokemonSerializer(serializers.ModelSerializer):
    class Meta:
        model = Pokemon
        fields = ['id', 'identifier']


class ItemSerializer(serializers.ModelSerializer):
    class Meta:
        model = Items
        fields = ['id', 'identifier']