requirements.txt
absl-py==0.12.0
aiohttp==3.7.4.post0
appdirs==1.4.4
asgiref==3.4.1
astunparse==1.6.3
async-timeout==3.0.1
asyncpg==0.23.0
attrs==21.2.0
backcall==0.2.0
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
clang==5.0
click==8.0.1
colorama==0.4.4
cycler==0.10.0
decorator==5.0.9
dill==0.3.3
discord.py==1.7.2
distlib==0.3.2
Django==3.2.8
django-filter==21.1
djangorestframework==3.12.4
emnist==0.0
enum34==1.1.10
filelock==3.0.12
flatbuffers==1.12
future==0.18.2
gast==0.4.0
google-auth==1.30.1
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
googleapis-common-protos==1.53.0
grpcio==1.39.0
h5py==3.1.0
idna==2.10
ipython==7.24.1
ipython-genutils==0.2.0
jedi==0.18.0
joblib==1.0.1
keras==2.6.0
Keras-Applications==1.0.8
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.4
matplotlib==3.4.2
matplotlib-inline==0.1.2
multidict==5.1.0
mysqlclient==2.0.3
numpy==1.19.5
oauthlib==3.1.1
opencv-contrib-python==4.5.2.52
opt-einsum==3.3.0
parso==0.8.2
pickleshare==0.7.5
Pillow==8.2.0
pipenv==2021.5.29
plaidbench==0.7.0
plaidml==0.7.0
plaidml-keras==0.7.0
playsound==1.2.2
promise==2.3
prompt-toolkit==3.0.18
protobuf==3.17.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.9.0
PyMySQL==1.0.2
pyparsing==2.4.7
PySimpleGUI==4.43.0
pysound==0.1
python-dateutil==2.8.1
pytz==2021.3
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.7.0
six==1.15.0
sklearn==0.0
sqlparse==0.4.2
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.6.0
tensorflow-datasets==4.3.0
tensorflow-estimator==2.6.0
tensorflow-examples===f8ea518e0efdbf5d7c75b314133b29e8e0212e70-
tensorflow-hub==0.12.0
tensorflow-metadata==1.0.0
tensorflow-text==2.6.0
tensorflowjs==3.8.0
termcolor==1.1.0
threadpoolctl==2.2.0
tqdm==4.61.0
traitlets==5.0.5
typing-extensions==3.7.4.3
urllib3==1.26.5
virtualenv==20.4.7
virtualenv-clone==0.5.4
wcwidth==0.2.5
Werkzeug==2.0.1
wrapt==1.12.1
yarl==1.6.3
